Fix another ordering bug with streams and shortenPath().
The LocalClient wrapper around a capability server implementation maintains a call queue when streaming. So if the object's `shortenPath()` indicates a shorter path, we have to actually wait for said call queue to drain out before we allow subsequent calls to go directly to the shorter path.

This is very similar to embargoes in the RPC system, but no RPC is necessarily involved here.